requirements.txt
streamlit
numpy
faiss-cpu
requests
sentence-transformers
PyPDF2
diffusers
transformers
accelerate
peft